% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDictionary.R
\name{DataDictionary}
\alias{DataDictionary}
\title{A Data Dictionary Function}
\usage{
DataDictionary(x, variable_labels, value_labels)
}
\arguments{
\item{x}{A dataframe}

\item{variable_labels}{A vector or list of variable labels. Should be the same length as ncol(x) and the labels should appear in the same order.}

\item{value_labels}{A list of value labels.}

\item{file}{The file name and location to which the data dictionary should be written. Defaults to "DataDictionary.txt" in the current working directory.}
}
\description{
This function creates a .txt data dictionary from a dataframe, list of variable labels, and list of value labels. It writes these things to a data dictionary file that can function as a codebook.
}
\examples{
DataDictionary()
}
\keyword{data}
\keyword{dictionary}
\keyword{metadata}

