% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DictionarySPSS.R
\name{DictionarySPSS}
\alias{DictionarySPSS}
\title{A SPSS Data Dictionary Function}
\usage{
DictionarySPSS(x, file = "DataDictionary.txt")
}
\arguments{
\item{x}{A dataframe read in using the "read.spss()" command from the foreign package. Must have "to.data.frame=TRUE" and works best if "use.value.labels=FALSE"}

\item{file}{The file name and location to which the data dictionary should be written. Defaults to "DataDictionary.txt" in the current working directory.}
}
\description{
This function creates a .txt data dictionary from an SPSS .sav file read into R using the foreign package. It writes the variable labels and value labels stored as attributes within the R data object to an external file.
}
\examples{
DictionarySPSS()
}
\keyword{data}
\keyword{dictionary}
\keyword{metadata}

