% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DictionaryStata.R
\name{DictionaryStata}
\alias{DictionaryStata}
\title{A Stata Data Dictionary Function}
\usage{
DictionaryStata(x, file = "DataDictionary.txt")
}
\arguments{
\item{x}{A dataframe read in using the "read.dta()" command from the foreign package.}

\item{file}{The file name and location to which the data dictionary should be written. Defaults to "DataDictionary.txt" in the current working directory.}
}
\description{
This function creates a .txt data dictionary from a Stata .dta read into R using the foreign package. It writes the variable labels and value labels stored as attributes within the R data object to an external file.
}
\examples{
DictionaryStata()
}
\keyword{data}
\keyword{dictionary}
\keyword{metadata}

