% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_stat.R
\name{mode_stat}
\alias{mode_stat}
\title{Calculate mode (statistic)}
\usage{
mode_stat(x, freq = FALSE)
}
\arguments{
\item{x}{A vector}

\item{freq}{If true the frequency of the mode will display.}
}
\value{
if freq = FALSE mode_stat will return a single value. If freq = TRUE it will return a dataframe
}
\description{
Calculate mode (statistic)
}
\examples{
foo <- sample(x = c("L1", "L2", "L3", "L4", "L5"), size = 200, replace = TRUE,
  prob = c(0.2, 0.4, 0.1, 0.05, 0.25))
bar <- floor(rnorm(n = 200, mean = 50, sd = 5))
dat <- as.data.frame(cbind(foo, bar))

mode_stat(dat$foo)

mode_stat(dat$bar, freq = TRUE)

}
