% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomID.R
\name{RandomID}
\alias{RandomID}
\title{A Random ID function}
\usage{
RandomID(x, identifier.variable, seed.value, id.digits)
}
\arguments{
\item{x}{}
}
\description{
This function replaces a list of identifiers with random numeric ids and replaces the identifying variable with a variable called "RandomID" in the dataframe.
}
