% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_variable.R
\name{summarise_variable}
\alias{summarise_variable}
\title{Summarise a single variable}
\usage{
summarise_variable(dataset, column)
}
\arguments{
\item{dataset}{The dataset with the variable you wish to summarise}

\item{column}{The column you wish to summarise as a quoted string}
}
\value{
A dataframe with a summary of the variable
}
\description{
Summarise a single variable
}
\examples{
 summarise_variable(mtcars, "mpg")

 summarise_variable(iris, "Species")
}
