% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dictionary.R
\name{create_dictionary}
\alias{create_dictionary}
\title{Create a data dictionary from any dataset}
\usage{
create_dictionary(dataset, id_var = NULL, file = NULL, ...)
}
\arguments{
\item{dataset}{The dataset you wish to summarise}

\item{id_var}{A variable or vector of variables that are identifiers (optional)}

\item{file}{The filepath the write an Excel spreadsheet (optional)}

\item{...}{Other arguments}
}
\value{
Either an Excel spreadsheet or a dataframe
}
\description{
Create a data dictionary from any dataset
}
\examples{
 create_dictionary(esoph)

 mtcars$id <- 1:nrow(mtcars)
 create_dictionary(mtcars, id_var = "id")

}
