% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dictionary.R
\name{create_dictionary}
\alias{create_dictionary}
\title{Create a data dictionary from any dataset}
\usage{
create_dictionary(dataset, id_var = NULL, file = NULL, output = FALSE, ...)
}
\arguments{
\item{dataset}{The dataset you wish to summarise}

\item{id_var}{A variable or vector of variables that are identifiers (optional)}

\item{file}{The filepath the write an Excel spreadsheet}

\item{output}{Output to console (default = FALSE)}

\item{...}{Other arguments}
}
\value{
Either an Excel spreadsheet or a dataframe
}
\description{
Note: you must specify one of output = TRUE or supply an argument to file.
}
