% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode_stat}
\alias{mode_stat}
\title{Get the mode of a vector}
\usage{
mode_stat(x, freq = FALSE)
}
\arguments{
\item{x}{A vector}

\item{freq}{Boolean when TRUE returns the frequency of the mode}
}
\description{
Get the mode of a vector
}
\keyword{internal}
